move SecretHash to bls and put plaintext back
