+++ crypto/bls/src/lib.rs
@@ -4,8 +4,8 @@ extern crate ssz;
 #[macro_use]
 mod macros;
 mod keypair;
-mod plain_text;
 mod public_key_bytes;
+mod secret_hash;
 mod secret_key;
 mod signature_bytes;
 mod signature_set;
@@ -14,7 +14,7 @@ pub use crate::keypair::Keypair;
 pub use crate::public_key_bytes::PublicKeyBytes;
 pub use crate::secret_key::SecretKey;
 pub use crate::signature_bytes::SignatureBytes;
-pub use plain_text::PlainText;
+pub use secret_hash::SecretHash;
 pub use signature_set::{verify_signature_sets, SignatureSet};
 
 #[cfg(feature = "arbitrary")]
