+++ crypto/eth2_key_derivation/src/plain_text.rs
