+++ crypto/eth2_keystore/src/lib.rs
@@ -6,7 +6,8 @@ mod keystore;
 
 pub mod json_keystore;
 
-pub use bls::PlainText;
+pub use bls::SecretHash;
+pub use eth2_key_derivation::PlainText;
 pub use keystore::{
     decrypt, default_kdf, encrypt, keypair_from_secret, Error, Keystore, KeystoreBuilder, DKLEN,
     HASH_SIZE, IV_SIZE, SALT_SIZE,
